// frontend/src/api/generated/models/FileRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Ответ с метаданными файла.
 */
export type FileRead = {
    id: string;
    filename: string;
    content_type: string;
    size_bytes: number;
    storage_key: string;
    is_public: boolean;
    created_at: string;
    updated_at: string;
};